Add a dummy non-empty environment when using assert_python_{failure,ok}
If the env_vars argument to assert_python_failure or assert_python_ok
is empty the test will run in isolated mode (-I) which means that the
PYTHONPATH environment variable will be ignored and the test fails
because the isal module can not be found, or the test is run usung the
system installed version of the module instead of the newly built
module that should be tested.

By adding a dummy entry to the env_vars argument the isolated mode is
not used and the PYTHONPATH environment variable is not ignored and
the test works as expected.